moved participant context events to "events" package
